API-12090 copy changes from v3.4
