+++ src/agent/web.ts
@@ -1,32 +1,36 @@
 import { WebAPI } from "../internal";
 import { TokenGetter } from "../authorization";
 import type {
-  ListChatsResponse,
-  ListChatParameters,
+  AgentForTransfer,
+  ChatAccess,
+  CreateCustomerResponse,
+  CustomerParameters,
   EmptyResponse,
-  ListThreadsParameters,
-  ListThreadsResponse,
+  Event,
   GetChatResponse,
+  GetCustomerResponse,
   ListArchivesParameters,
   ListArchivesResponse,
-  StartChatParameters,
-  StartChatResponse,
-  ResumeChatParameters,
-  ResumeChatResponse,
-  TransferChatParameters,
-  SendEventResponse,
-  SendRichMessagePostbackParameters,
-  GetCustomerResponse,
+  ListChatParameters,
+  ListChatsResponse,
   ListCustomersParameters,
   ListCustomersResponse,
-  CustomerParameters,
-  CreateCustomerResponse,
+  ListThreadsParameters,
+  ListThreadsResponse,
   MulticastRecipients,
-  AgentForTransfer,
-  UploadFileResponse,
+  Properties,
+  ResumeChatParameters,
+  ResumeChatResponse,
+  RoutingStatus,
+  SendEventResponse,
+  SendRichMessagePostbackParameters,
   SetRoutingStatusResponse,
+  StartChatParameters,
+  StartChatResponse,
+  TransferChatParameters,
+  UploadFileResponse,
+  WebAPIOptions,
 } from "./structures";
-import { ChatAccess, Event, Properties, RoutingStatus, WebAPIOptions } from "../objects";
 import { promises as fs } from "fs";
 import axios from "axios";
 import FormData from "form-data";
@@ -163,6 +167,7 @@ export default class Web extends WebAPI {
       require_active_thread,
     });
   }
+
   /**
    * Removes a user from chat. Removing customer user type is not allowed. It's always possible to remove the requester from the chat.
    * @param chat_id - chat to remove user from
@@ -175,6 +180,7 @@ export default class Web extends WebAPI {
       user_type: "agent",
     });
   }
+
   /**
    * Sends an Event object. Use this method to send a message by specifing the Message event type in the request.
    * It's possible to write to a chat without joining it. The user sending an event will be automatically added to the chat
