--- src/configuration/structures.js
@@ -1,10 +1,0 @@
-"use strict";
-exports.__esModule = true;
-exports.GroupPriority = void 0;
-var GroupPriority;
-(function (GroupPriority) {
-    GroupPriority["First"] = "first";
-    GroupPriority["Normal"] = "normal";
-    GroupPriority["Last"] = "last";
-    GroupPriority["DoNotAssign"] = "supervisor";
-})(GroupPriority = exports.GroupPriority || (exports.GroupPriority = {}));
