+++ src/configuration/structures/index.ts
@@ -1,0 +1,2 @@
+export * from "./responses";
+export * from "./structures";
