+++ src/configuration/structures/responses.ts
@@ -1,0 +1,29 @@
+export interface EmptyResponse {}
+
+export interface CreateAgentResponse {
+  id: string;
+}
+
+export interface CreateBotResponse {
+  id: string;
+}
+
+export interface CreateGroupResponse {
+  id: string;
+}
+
+export interface RegisterWebhookResponse {
+  id: string;
+}
+
+export interface AddAutoAccessResponse {
+  id: string;
+}
+
+export interface GetOrganizationIDResponse {
+  organization_id: string;
+}
+
+export interface GetLicenseIDResponse {
+  license_id: string;
+}
