--- src/internal/constants.js
@@ -1,5 +1,0 @@
-"use strict";
-exports.__esModule = true;
-exports.ApiURL = exports.ApiVersion = void 0;
-exports.ApiVersion = "3.3";
-exports.ApiURL = "api.livechatinc.com";
