--- src/webhooks/index.js
@@ -1,2 +1,0 @@
-"use strict";
-exports.__esModule = true;
